refactor: Aplply suggestions
