+++ src/core/moenda.js
@@ -1,16 +1,14 @@
 const path = require('path');
 const {translateOptions} = require(path.resolve(__dirname, '../loader'));
-const Linter = require(path.resolve(__dirname, './linter.js'));
+const {parseCommentSuppression} = require(path.resolve(
+  __dirname,
+  '../utils/parserUtils.js',
+));
 
 class Moenda {
   constructor(options) {
     this.results = [];
     this.config = translateOptions(options);
-    this.linter = new Linter(
-      this.config.rules,
-      this.config.processor,
-      this.config.parser,
-    );
   }
 
   reporter(ruleName, filePath) {
@@ -24,11 +22,27 @@ class Moenda {
     };
   }
 
+  verify(file, report, config) {
+    const tokens = this.config.parser.parse(file.content);
+    const context = this.config.processor(tokens);
+    const params = {context, config};
+
+    const disabledRules = parseCommentSuppression(tokens);
+    const enabledRules = this.config.rules.filter(
+      (rule) => !disabledRules.hasOwnProperty(rule.name),
+    );
+
+    enabledRules.forEach((rule) => {
+      rule.run(params, report(rule.name, file.path));
+    });
+  }
+
   runRules() {
     const {files, rulesConfig} = this.config;
     const reporter = this.reporter.bind(this);
-    files.forEach((file) => this.linter.verify(file, reporter, rulesConfig));
+    files.forEach((file) => this.verify(file, reporter, rulesConfig));
   }
+
 
   getResults() {
     return this.results;
